// src/backend/shared/enums/error-codes.enum.ts
export enum UserErrorCode {
    USER_NOT_FOUND = 'USER_NOT_FOUND',
    EMAIL_ALREADY_EXISTS = 'EMAIL_ALREADY_EXISTS',
    USERNAME_ALREADY_EXISTS = 'USERNAME_ALREADY_EXISTS',
    INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',
    INVALID_PASSWORD = 'INVALID_PASSWORD',
    INVALID_EMAIL_FORMAT = 'INVALID_EMAIL_FORMAT',
    INVALID_USERNAME_FORMAT = 'INVALID_USERNAME_FORMAT'
}